pyproject.toml
[project]
name = "tickertape"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "edgartools>=4.20.0",
    "fastapi>=0.119.0",
    "openai-agents>=0.3.3",
    "tqdm>=4.67.1",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.19.0",
]
